// src/renderer/src/components/inventarioYlogistica/historiales/contenedores/types/types.d.ts
/* eslint-disable prettier/prettier */
export type itemListaEmpaqueType = {
    pallet: string
    fecha: string
    label: string
    variedad: string
    producto: string
    peso: string
    categoria: number
    tamaño: number
    cantidad: number
    ICA: string
    GGN: string
    fechaVencimiento: string
}